+++ apps/web/src/utils/formatUsdPrice.ts
@@ -1,0 +1,8 @@
+import { formatEther } from 'viem';
+
+export function formatUsdPrice(price: bigint, ethUsdPrice: number) {
+  if (price === 0n) return '0';
+  const parsed = (parseFloat(formatEther(price)) * Number(ethUsdPrice)).toFixed(2);
+  if (parsed === '0.00') return '0';
+  return parsed;
+}
